#136: Move RelationalDb functionality into a facade class for easier reuse
