+++ broker/src/functional-test/groovy/com/swisscom/cloud/sb/broker/functional/BaseFunctionalSpec.groovy
@@ -26,6 +26,7 @@ import org.joda.time.LocalTime
 import org.joda.time.Seconds
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.http.client.HttpComponentsClientHttpRequestFactory
 import org.springframework.test.context.ContextConfiguration
 import org.springframework.web.client.RestTemplate
 import org.springframework.web.context.WebApplicationContext
@@ -100,5 +101,9 @@ abstract class BaseFunctionalSpec extends Specification {
                 Thread.sleep(sleepTime)
             }
         }
+    }
+
+    protected String readResourceContent(String resourcePath) {
+        return new File(getClass().getResource(resourcePath).getFile()).text
     }
 }
